python/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
bitarray==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.2
cytoolz==0.12.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
flashbots==1.1.1
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.9.0
parsimonious==0.8.1
protobuf==3.19.5
pycryptodome==3.19.0
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==2.0.1
rpds-py==0.10.6
six==1.16.0
toolz==0.12.0
typing_extensions==4.8.0
urllib3==2.0.7
varint==1.0.2
web3==5.31.4
websockets==9.1
yarl==1.9.2
